auth/requirements.txt
Flask==3.0.3
# psycopg2-binary==2.9.9
mysql-connector-python==8.4.0
requests==2.25.1